requirements.txt
alembic==1.7.4
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
fake-useragent==0.1.11
Flask==2.0.1
Flask-Caching==1.10.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
Mako==1.1.5
MarkupSafe==2.0.1
parse==1.19.0
psycopg2==2.9.1
pyee==8.1.0
pyppeteer==0.2.5
pyquery==1.4.3
python-dotenv==0.19.1
requests==2.25.1
requests-cache==0.6.4
requests-html==0.10.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.26
tqdm==4.61.1
url-normalize==1.4.3
urllib3==1.26.6
w3lib==1.22.0
websockets==8.1
Werkzeug==2.0.1
